Add symbols for proto-compute, and fill out symbols for can-compute; break out proto-compute tests to their own file
